feat: add move cars
