+++ src/racingcar/store/carsStore.js
@@ -1,4 +1,5 @@
 import Car from '../models/Car';
+import { randomNumber } from '../../utils/randoms';
 
 const carsStore = {
   car_names: [],
@@ -9,7 +10,7 @@ const carsStore = {
   SET_CAR_NAMES: (_car_names) => (carsStore.car_names = _car_names),
   GET_CARS: () => carsStore.cars,
   SET_CARS: (_car_names) => (carsStore.cars = _car_names.map((name) => new Car(name))),
-  MOVE_CARS: () => carsStore.GET_CARS().forEach((car) => car.move()),
+  MOVE_CARS: () => carsStore.GET_CARS().forEach((car) => car.move(randomNumber())),
   GET_WINNERS: () => carsStore.winners,
   FIND_AND_SET_WINNERS: () =>
     (carsStore.winners = carsStore
